==> crawling_code/01_crawling_melon_chart(1).ipynb <==
import requests
from bs4 import BeautifulSoup
from fake_useragent import UserAgent

### step by step process (get response, BeautifulSoup, select through css, put title-artist-link data in dataframe)

year = 2013
url = "https://www.melon.com/chart/age/list.htm?idx=1&chartType=YE&chartGenre=KPOP&chartDate={}&moved=Y".format(year)
headers = {
    "User-Agent" : UserAgent().chrome
}

response = requests.get(url, headers=headers)
response

dom = BeautifulSoup(response.text, 'html.parser')
elements = dom.select("#frm > table > tbody tr")
len(elements)

element = elements[0]

data = {
    "title" : element.select_one('a > .odd_span').text.split('상')[0].strip(),
    "artist" : element.select_one('div > .fc_mgray').text,
    "link" : 'https://www.melon.com/song/detail.htm?songId=' + element.select_one('td > div > input').get('value')
}

datas = []
for element in elements:
    data = {
        "title" : element.select_one('a > .odd_span').text.split('상')[0].strip(),
        "artist" : element.select_one('div > .fc_mgray').text,
        "link" : 'https://www.melon.com/song/detail.htm?songId=' + element.select_one('td > div > input').get('value')
    }
    datas.append(data)

import pandas as pd

song_df = pd.DataFrame(datas)
song_df.tail()

### get genre data of each song through link column

link = song_df.loc[0, 'link']
print(link)

headers = {
    "User-Agent" : UserAgent().chrome
}
response = requests.get(link, headers=headers)

dom = BeautifulSoup(response.content, 'html.parser')
response

elements = dom.select_one('div.entry > div.meta > dl > dd:nth-child(6)').text
elements

# 장르 수집 함수
# 한꺼번에 너무 많이 수집하면 도중에 response 406이 발생하여 상세 페이지 접속이 불가함
# 년도를 나눠서 수집하면 정상적으로 수집할 수 있으나 한 번 실행하고 나면 response 406이 발생하기 때문에 한시간 정도 쉬어주면서 다시 실행해야 함

def get_genre(link):
    headers = {"User-Agent" : UserAgent().chrome}
    response = requests.get(link, headers=headers)
    dom = BeautifulSoup(response.content, 'html.parser')
    
    return dom.select_one('div.entry > div.meta > dl > dd:nth-child(6)').text

# 시대별 차트 수집 함수
# 문제 없이 정상적으로 작동

def song_details(year):

        url = "https://www.melon.com/chart/age/list.htm?idx=1&chartType=YE&chartGenre=KPOP&chartDate={}&moved=Y".format(year)
        headers = {
            "User-Agent" : UserAgent().chrome
        }
        response = requests.get(url, headers=headers)
        dom = BeautifulSoup(response.text, 'html.parser')
        elements = dom.select("#frm > table > tbody tr")
        
        datas = []
        for element in elements:
            data = {
                "Title" : element.select_one('a > .odd_span').text.split('상')[0].strip(),
                "Artist" : element.select_one('div > .fc_mgray').text,
                "Year" : year,
                "Link" : 'https://www.melon.com/song/detail.htm?songId=' + element.select_one('td > div > input').get('value')
            }
            datas.append(data)
    
        return datas

### 1964~1975

year = list(range(1964, 1976))

result = []
for y in year:
    result.extend(song_details(y))

first_df = pd.DataFrame(result)

first_df['Genre'] = first_df['Link'].apply(get_genre)
first_df.tail()

first_df.to_csv('./datas/first.csv')

### 1976~1985

year = list(range(1976, 1986))

result = []
for y in year:
    result.extend(song_details(y))

second_df = pd.DataFrame(result)
second_df.tail()

second_df['Genre'] = second_df['Link'].apply(get_genre)
second_df.tail()

second_df.to_csv('./datas/second.csv')

### 1986~1995

year = list(range(1986, 1996))

result = []
for y in year:
    result.extend(song_details(y))

third_df = pd.DataFrame(result)
third_df.tail()

third_df['Genre'] = third_df['Link'].apply(get_genre)
third_df.tail()

third_df.to_csv('./datas/third.csv')

### 2006~2010

year = list(range(2006, 2011))

result = []
for y in year:
    result.extend(song_details(y))

fourth_df = pd.DataFrame(result)
fourth_df.tail()

fourth_df['Genre'] = fourth_df['Link'].apply(get_genre)
fourth_df.tail()

fourth_df.to_csv('./datas/sixth.csv')

### 2016~2019

year = list(range(2016, 2020))

result = []
for y in year:
    result.extend(song_details(y))

fifth_df = pd.DataFrame(result)
fifth_df.tail()

fifth_df['Genre'] = fifth_df['Link'].apply(get_genre)
fifth_df.tail()

fifth_df.to_csv('./datas/eigth.csv')
