==> crawling_code/03_concat_separate_dataframe.ipynb <==
### concat separate dataframes
### eight dataframes were made while crawling 'Genre' column because crawling large data blocked us from the page so we had to separate the years to bring them all

import pandas as pd
import seaborn as sns
import matplotlib.pyplot as plt
%matplotlib inline

first = pd.read_csv('./datas/first.csv', index_col=[0])
first.tail()

second = pd.read_csv('./datas/second.csv', index_col=[0])
second.tail()

third = pd.read_csv('./datas/third.csv', index_col=[0])
third.tail()

fourth = pd.read_csv('./datas/fourth.csv', index_col=[0])
fourth.tail()

frames = [first, second, third, fourth]
df = pd.concat(frames)
df.drop(columns=['Link'], inplace=True)

df.to_csv('./datas/df~2000.csv')

fifth = pd.read_csv('./datas/fifth.csv', index_col=[0])
fifth.tail()

frame = [df, fifth]
df = pd.concat(frame)
df

sixth = pd.read_csv('./datas/sixth.csv', index_col=[0])
sixth.tail()

seventh = pd.read_csv('./datas/seventh.csv', index_col=[0])
seventh.tail()

frames = [df, sixth, seventh]
df = pd.concat(frames)
df.tail()

df

df.to_csv('./datas/df~2015.csv')

eigth = pd.read_csv('./datas/eigth.csv', index_col=[0])
eigth.tail()

frames = [df, eigth]
df = pd.concat(frames)
df.drop(columns=['Link'], inplace=True)

df.reset_index(drop=True, inplace=True)
df

df.to_csv('./datas/final2019.csv')
