==> crawling_code/04_rename_genre.ipynb <==
df = pd.read_csv('C:/Users/luvu1/Documents/dss15/crawling_project/datas/final2019(rev3).csv', index_col=[0])

df.tail()

### some genres were paired or grouped together, some were listed with just one genre
### renamed and replaced the values of 'Genre' column to limit the genre kinds to only 'R&B/Soul', '국내영화/국내드라마', '그외', '댄스', '랩/힙합', '록/메탈', '발라드', '성인가요', '인디음악', '재즈', '포크/블루스'

df.groupby('Genre').size().to_frame()

df = df[df.Genre != "-"]
df = df[df.Genre != "POP"]

df.drop_duplicates(inplace=True)

df = df.replace(["랩/힙합, 국내드라마", "포크/블루스, 국내드라마", "발라드, 랩/힙합, 국내드라마", "발라드, 국내드라마", "발라드, R&B/Soul, 국내드라마", "록/메탈, 국내드라마", "국내드라마", "록/메탈, 국내영화", "발라드, 국내영화", "일렉트로니카, 국내드라마", "국내영화", "댄스, 일렉트로니카"], "국내영화/국내드라마")

df = df.replace(["R&B/Soul, 인디음악", "랩/힙합, 인디음악", "인디음악, 록/메탈", "인디음악, 일렉트로니카", "인디음악, 포크/블루스", "발라드, 인디음악"], "인디음악")

df = df.replace(["댄스, 성인가요", "발라드, 성인가요"], "성인가요")

df = df.replace(["국악, 국악가요", "뉴에이지", "재즈", "일렉트로니카"], "그외")

df = df.replace(["록/메탈, 포크/블루스", "발라드, 포크/블루스"], "포크/블루스")

df = df.replace(["댄스, R&B/Soul", "발라드, 댄스, 랩/힙합"], "댄스")

df = df.replace(["발라드, R&B/Soul"], ["R&B/Soul"])

df = df.replace(["발라드, 랩/힙합", "댄스, 랩/힙합"], "랩/힙합")

df = df.replace(["발라드, 록/메탈"], ["록/메탈"])

df.loc[df["Title"] == "G Cafe", 'Genre'] = "댄스"
df.loc[df["Title"] == "너의 궁전", 'Genre'] = "댄스"
df.loc[df["Title"] == "다시", 'Genre'] = "발라드"
df.loc[df["Title"] == "부탁", 'Genre'] = "발라드"

a = df[df["Title"] == "연정/폼생폼사/기사도"].index
df.drop(a, inplace=True)

df = df.replace(["재즈, 보컬재즈", "재즈, 애시드/퓨전/팝"], "재즈")

df.reset_index(drop=True, inplace=True)
df

df.to_csv('./datas/final2019(rev4).csv')
