==> crawling_code/07_write_module.ipynb <==
import pandas as pd

df = pd.read_csv('./datas/final2019(rev4).csv', index_col=[0])
df.tail()

# 모듈 생성

### had to manually make folder because of the following error

!rm -rf song
!mkdir song/top_music
!mkdir song/user
!touch song/song.py
!touch song/top_music/details.py
!touch song/user/user_playlist.py

!tree song

# 1. 음악 정보 수집 함수
### Melon TOP100 songs of each year between 1964~2019 (song Title, Artist, Link, Year)

%%writefile song/top_music/details.py

import requests
from bs4 import BeautifulSoup
from fake_useragent import UserAgent
import pandas as pd

def song_details(year):
    url = "https://www.melon.com/chart/age/list.htm?idx=1&chartType=YE&chartGenre=KPOP&chartDate={}&moved=Y".format(year)
    headers = {
        "User-Agent" : UserAgent().chrome
    }
    response = requests.get(url, headers=headers)
    dom = BeautifulSoup(response.text, 'html.parser')
    elements = dom.select("#frm > table > tbody tr")
    
    datas = []
    for element in elements:
        data = {
            "Title" : element.select_one('a > .odd_span').text.split('상')[0].strip(),
            "Artist" : element.select_one('div > .fc_mgray').text,
            "Year" : year,
            "Link" : 'https://www.melon.com/song/detail.htm?songId=' + element.select_one('td > div > input').get('value')
        }
        datas.append(data)
    df = pd.DataFrame(datas)
    
    return df

#상세 페이지 내 장르 수집 함수
def get_genre(link):
    headers = {"User-Agent" : UserAgent().chrome}
    response = requests.get(link, headers=headers)
    dom = BeautifulSoup(response.content, 'html.parser')
    
    return dom.select_one('dd:nth-of-type(3)').text

#상세 페이지 내 가사 수집 함수
def get_lyrics(link):
    headers = {"User-Agent" : UserAgent().chrome}
    response = requests.get(link, headers=headers)
    dom = BeautifulSoup(response.content, 'html.parser')
    elements = dom.select_one('.wrap_lyric').text.replace('\r', '').replace('\n', '').replace('\t', '').split('펼치기')[0]
    if elements == '[가사 준비중] 멜론 회원 여러분! 가사 등록을 기다리고 있어요.가사등록하기':
        elements = '-'
        
    return elements

def run(year):
    df = song_details(year)
    df['Genre'] = df['Link'].apply(get_genre)
    df['Lyrics'] = df['Link'].apply(get_lyrics)
    df.drop(columns=['Link'], inplace=True)
    
    return df

# 모듈 작동 확인

import song.libs.details as details

details.run(1964)

# 2. 사용자 플레이리스트 수집 함수
### only melon users' playlist can be gathered

%%writefile song/user/user_playlist.py

import requests
from bs4 import BeautifulSoup
from fake_useragent import UserAgent
import pandas as pd

def playlist(seq):
    url = "https://www.melon.com/mymusic/playlist/mymusicplaylistview_listSong.htm?plylstSeq={}".format(seq)
    headers = {
        "User-Agent" : UserAgent().chrome
    }
    response = requests.get(url, headers=headers)
    dom = BeautifulSoup(response.text, 'html.parser')
    elements = dom.select("#frm > div > table > tbody tr")
    
    datas = []
    for element in elements:
        data = {
            "Title" : element.select_one("a.fc_gray").text,
            "Artist" : element.select_one("#artistName").text.replace('\n', ''),
            "Link" : 'https://www.melon.com/song/detail.htm?songId=' + element.select_one('td > div > input').get('value')
            }
        datas.append(data)
    df_user = pd.DataFrame(datas)
    
    return df_user

#상세 페이지 내 장르 수집 함수
def get_genre(link):
    headers = {"User-Agent" : UserAgent().chrome}
    response = requests.get(link, headers=headers)
    dom = BeautifulSoup(response.content, 'html.parser')
    
    return dom.select_one('dd:nth-of-type(3)').text

#상세 페이지 내 가사 수집 함수
def get_lyrics(link):
    headers = {"User-Agent" : UserAgent().chrome}
    response = requests.get(link, headers=headers)
    dom = BeautifulSoup(response.content, 'html.parser')
    elements = dom.select_one('.wrap_lyric').text.replace('\r', '').replace('\n', '').replace('\t', '').split('펼치기')[0]
    if elements == '[가사 준비중] 멜론 회원 여러분! 가사 등록을 기다리고 있어요.가사등록하기':
        elements = '-'
        
    return elements

def user(seq):
    df_user = playlist(seq)
    df_user['Genre'] = df_user['Link'].apply(get_genre)
    df_user['Lyrics'] = df_user['Link'].apply(get_lyrics)
    df_user.drop(columns=['Link'], inplace=True)
    
    return df_user

# 모듈 작동 확인

import song.user.user_playlist as playlist

playlist.user(486216850)

!mkdir song/recomm
!tree song

# 3. 가사 분석 후 사용자가 입력한 단어와 유사한 음악 추천 함수

%%writefile song/recomm/lyrics_analysis.py

from konlpy.tag import Okt
t = Okt()
import re
import gensim
from gensim.models import word2vec
import pandas as pd
import pymongo

client = pymongo.MongoClient('mongodb://dss:dss@'ip':27017')
db = client['melon_chart']
collection = db.song_list
df = pd.DataFrame(list(collection.find()))
df.drop(columns=['_id'], inplace=True)

# 가사 전처리: 한글, 영어, 숫자만 남기고 모든 특수문자 제거
def cleanse(text):
    pattern = re.compile(r'\s+')
    text = re.sub(pattern, ' ', text)
    text = re.sub('[^가-힣ㄱ-ㅎㅏ-ㅣa-zA-Z0-9]', ' ', text)
    return text

# 기간, 연관 단어 입력 하여 유사한 음악 추첯
def search_similarity(year_end, year_start, look_keyword):
    yrange = df[(df['Year'] < year_end) & (df['Year'] >= year_start)].reset_index(drop=True)
    lyrics = yrange['Lyrics'].apply(cleanse)
    lyrics = lyrics.tolist()
        
    results = []
    for i in lyrics:
        result = t.pos(i, norm=True, stem=True)
        
        word = []
        for a in result:
            if not a[1] in ["Josa", "Determiner", "Exclamation", "Suffix"]:
                word.append(a[0])
        words = (" ".join(word)).strip()
        results.append(words)
        
    data_file = 'lyrics.data'
    with open(data_file, 'w', encoding='utf-8') as fp:
        fp.write("\n".join(results))
    data = word2vec.LineSentence(data_file)
    model = word2vec.Word2Vec(data, size=200, window=10, hs=1, min_count=2, sg=1)
    model.save('lyrics.model')
    model = word2vec.Word2Vec.load("lyrics.model")
    song = model.wv.most_similar(look_keyword, topn=10)
    keyword = []
    for s in song:
        keyword.append(s[0])
    relist = yrange[yrange['Lyrics'].str.contains('|'.join(keyword))]
    
    return relist.sample(n=10)

# 모듈 작동 확인

import song.recomm.lyrics_analysis as search

dj = search.search_similarity(2015, 2010, '사랑')

dj.to_dict(orient="recode")

dj1 = dj[['Artist', 'Title', 'Genre']].values
dj1

# 4. 음악 데이터와 사용자 플레이리스트간 장르 유사도 측정 후 음악 추천 함수

%%writefile song/recomm/ranking_random.py

import pandas as pd
import pymongo
import math
import numpy as np
import random
import song.user.user_playlist as playlist

# mongodb에 연결해서 dataframe 만들기 : 전체곡 정보 데이터프레임 : df
client = pymongo.MongoClient('mongodb://dss:dss@'ip':27017')
db = client['melon_chart']
collection = db.song_list
df = pd.DataFrame(list(collection.find()))
df.drop(columns=['_id'], inplace=True)


# 장르선호도가 유사한 연도 ranking top5 뽑는 함수
def top5_list(seq, df=df):
    # NEW DATAFRAME
    recommend = []
    recommend_df = pd.DataFrame()
    
    # 1. 전체인기곡들 장르유사도
    df = df.reset_index(drop = True)
    df["count"] = 1
    df_pivot = pd.pivot_table(data = df , values = "count", index = "Year", columns = "Genre",aggfunc = np.sum)
    df_pivot.fillna(0, inplace = True)
    df_normal = round(df_pivot.div(df_pivot.sum(axis = 1), axis=0),3)
    df_normal = df_normal.reset_index()
    
    # 2. 유저플레이리스트 장르유사도  함수
    df_user = playlist.user(seq)
    df_user.reset_index()
    # PIVOT
    df_user["count"] = 1
    user_pivot = pd.pivot_table(data = df_user, values = "count", columns="Genre",  aggfunc = np.sum)
    user_pivot.fillna(0, inplace=True)
    # 특정장르곡수/전체장르곡수 나눈 값 df : 사용자플레이리스트의 장르유사도 df
    user_normal = round(user_pivot.div(user_pivot.sum(axis=1), axis=0), 3)
    user_normal = pd.DataFrame(user_normal, columns=['R&B/Soul', '국내영화/국내드라마', '그외', '댄스','랩/힙합','록/메탈','발라드','성인가요','인디음악','재즈','포크/블루스'])
    np.nan_to_num(user_normal, copy=False)

    # 3. 장르유사도 비교 : 전체인기곡들 & 유저플레이리스트
    # 오차 구하기
    for i in range(len(df_normal)):
        error = (df_normal.loc[i] - user_normal) ** 2
        error = error.sum(axis =1)
        error = np.sqrt(error)
        recommend.append(error)
        
    # rms 칼럼만 만들기, 컬럼명 count를 rms로 변경
    from pandas import DataFrame
    recommend_df = pd.DataFrame(recommend)
    recommend_df.rename(columns={"count":"rms"}, inplace=True)
    
    # recommend_df 에 Year 값 넣은 df 만들기, df_normal에서 Year 컬럼 선택
    year = pd.DataFrame(df_normal["Year"])
    
    # Year 컬럼있는 df와 rms 컬럼있는 df 합치기
    recom = pd.concat([year,recommend_df], axis=1, join="inner")
    
    # 유사도 상위5개 : 인덱스 수정하기 위해 .reset_index(drop=True)
    ranking = recom.sort_values(by="rms", ascending=True).reset_index(drop=True)
    
    # 상위 5개 뽑기
    top5_r = pd.DataFrame()
    for i in range(0, 5):
        top5 = df[df["Year"] == ranking["Year"][i]]
        top5_r = pd.concat([top5_r, top5])

    fir = df[df["Year"] == ranking["Year"][0]] 
    sec = df[df["Year"] == ranking["Year"][1]]
    thr = df[df["Year"] == ranking["Year"][2]]
    four = df[df["Year"] == ranking["Year"][3]]
    fif = df[df["Year"] == ranking["Year"][4]]
    
    # 상위 5개 연도에서 가중치부여해서 random 곡 뽑기
    first = fir.take(np.random.permutation(len(fir))[:10])
    second = sec.take(np.random.permutation(len(fir))[:8])
    third = thr.take(np.random.permutation(len(fir))[:6])
    fourth = four.take(np.random.permutation(len(fir))[:4])
    fifth = fif.take(np.random.permutation(len(fir))[:2])
    
    # 뽑은 곡들 합해서 새로운 playlist 만들기
    result = pd.concat([first, second, third, fourth, fifth])
    
    # 곡 정보 중에서 가수명,곡명,장르명 만 보여주기
    playlist30 = result.loc[:, ["Artist", "Title", "Genre"]]
    
    return playlist30

webhook = "SLACK_WEBHOOK"

# 8. 슬랙 전송
### Sending to Slack channel through webhook 

%%writefile song/user/send_message.py

import requests
import json

def send_msg(webhook, dj, channel='#dss', username='DJ'):
    dj = str(dj[['Artist', 'Title', 'Genre']].values)
    payload = {"channel": channel, "username": username, "text": dj}
    requests.post(webhook, data=json.dumps(payload))

def send_msg(webhook, dj, channel='#dss', username='DJ'):
    dj = str(dj[['Artist', 'Title', 'Genre']].values)
    payload = {"channel": channel, "username": username, "text": dj}
    requests.post(webhook, data=json.dumps(payload))

send_msg(webhook, dj)

# 모듈 작동 확인

import song.user.slack as slack

slack.send_msg(webhook, dj)
