==> crawling_code/08_calculate_percent.ipynb <==
import pandas as pd

df = pd.read_csv("C:/Users/luvu1/Documents/dss15/crawling_project/datas/final2019(rev4).csv", encoding='utf-8', index_col=[0])
df

# select year 
# 1980 ~ 
df = df[df["Year"] >= 1980].reset_index(drop = True)

# make year_range

## elements 
bins = round((np.max(df["Year"]) - np.min(df["Year"]))/4)
labels = []
for i in (range(np.min(df["Year"]) , np.max(df["Year"]) +1,4)):
    labels.append(str(i) + '-' + str(i+3))


# Divided by  four-year cycles
## ex) 1980-1983 , ... ,2016-2019
df['year_range'] = pd.cut(x=df['Year'], bins= bins ,labels=labels)

# Count by genre

# 1. make count columns
df["count"] = 1

# pivoting!!
df_pivot = pd.pivot_table(data = df , values = "count", index = "year_range", columns = "Genre",aggfunc = np.sum)

# fillna --> 0
df_pivot.fillna(0, inplace = True)

# normalize

df_normal = round(df_pivot.div(df_pivot.sum(axis = 1), axis=0),3)
df_normal
