==> crawling_code/10_yj_playlist_data.ipynb <==
df = pd.read_csv('./datas/yj_list.csv', index_col=[0])

df.drop(columns=['Link'], inplace=True)
df

### data cleansing user playlist

df.groupby('Genre').size().to_frame()

df = df.replace(dict.fromkeys(["R&B;/Soul"], "R&B/Soul"))

df = df.replace(dict.fromkeys(["댄스, 국내드라마","발라드, 국내드라마", "발라드, 랩/힙합, 국내드라마"], "국내영화/국내드라마"))

df = df.replace(dict.fromkeys(["인디음악, 포크/블루스","R&B;/Soul, 인디음악", "인디음악, 록/메탈", "발라드, 인디음악"], "인디음악"))

df["count"] = 1
df_pivot = pd.pivot_table(data = df , values = "count", columns = "Genre",aggfunc = np.sum)
df_pivot.fillna(0, inplace = True)
df_pivot

df_normal = round(df_pivot.div(df_pivot.sum(axis = 1), axis=0),3)
df_normal
